feat: add invoice system

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 const { validationResult } = require('express-validator');
 
 exports.getAddProduct = (req, res, next) => {
@@ -14,20 +15,36 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
     const errors = validationResult(req);
-
-    if(!errors.isEmpty()){
-        return res.render('admin/edit-products', {
+    if (!image) {
+        return res.status(422).render('admin/edit-products', {
             pageTitle: 'Adicionar Produto',
-            path: '/admin/edit-products',
+            path: '/admin/add-products',
             editing: false,
             hasError: true,
             product: {
                 title: title,
-                imageUrl: imageUrl,
+                price: price,
+                description: description
+            },
+            errorMessage: 'O arquivo não é uma imagem.',
+            validationErros: []
+        });
+    }
+
+    const imageUrl = image.path; 
+
+    if(!errors.isEmpty()){
+        return res.status(422).render('admin/edit-products', {
+            pageTitle: 'Adicionar Produto',
+            path: '/admin/add-products',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
                 price: price,
                 description: description
             },
@@ -43,7 +60,7 @@ exports.postAddProduct = (req, res, next) => {
             res.redirect('/admin/products');
         })
         .catch(err => {
-            console.log(err);
+            res.redirect('/500');
         });
 };
 
@@ -75,10 +92,9 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedPrice = req.body.price;
     const updatedDescription = req.body.description;
-
     const errors = validationResult(req);
 
     if(!errors.isEmpty()){
@@ -89,7 +105,6 @@ exports.postEditProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: updatedTitle,
-                imageUrl: updatedImageUrl,
                 price: updatedPrice,
                 description: updatedDescription,
                 _id: prodId
@@ -107,7 +122,10 @@ exports.postEditProduct = (req, res, next) => {
         product.title = updatedTitle;
         product.price = updatedPrice;
         product.description = updatedDescription;
-        product.imageUrl = updatedImageUrl;
+        if (image) {
+            fileHelper.deleteFile(product.imageUrl);
+            product.imageUrl = image.path;
+        }
         return product.save().then(result => {
             res.redirect('/admin/products');
         })
@@ -117,12 +135,17 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
-        .then(() => {
-            res.redirect('/admin/products');
-        })
-        .catch(err => console.log(err));
+    Product.findById(prodId)
+    .then(product => {
+        if (!product){
+            return next(new Error('Produto não encontrado'));
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        return  Product.deleteOne({ _id: prodId, userId: req.user._id })
+    }).then(() => {
+        res.redirect('/admin/products');
+    })
+    .catch(err => console.log(err));
 }
 
 exports.adminProducts = (req, res, next) => {
